Updated documentation - Splitted error groups in different files - Added Introspection and Data Proxy errors - Updated missing errors
